chore: move Greeting dto to new module demo-util
