+++ client-jetty/src/main/java/module-info.java
@@ -1,6 +1,8 @@
 @SuppressWarnings({"deprecation", "removal"})
     module de.holisticon.servlet4demo.jettyclient {
 
+  requires de.holisticon.servlet4demo;
+
   requires spring.core;
   requires spring.beans;
   requires spring.context;
