+++ client-okhttp/pom.xml
@@ -7,10 +7,10 @@
     </parent>
     <modelVersion>4.0.0</modelVersion>
 
-    <artifactId>servlet4-demo-spring-boot-client</artifactId>
+    <artifactId>servlet4-demo-client-okhttp</artifactId>
     <packaging>jar</packaging>
 
-    <name>servlet4-demo-spring-boot-client</name>
+    <name>servlet4-demo-client-okhttp</name>
     <url>http://maven.apache.org</url>
 
     <properties>
@@ -43,6 +43,10 @@
         </plugins>
     </build>
     <dependencies>
+        <dependency>
+            <groupId>de.holisticon</groupId>
+            <artifactId>servlet4-demo-util</artifactId>
+        </dependency>
         <dependency>
             <groupId>org.springframework.boot</groupId>
             <artifactId>spring-boot-starter</artifactId>
