+++ client-okhttp/src/main/java/de/holisticon/servlet4demo/okhttpclient/Application.java
@@ -1,6 +1,6 @@
 package de.holisticon.servlet4demo.okhttpclient;
 
-import de.holisticon.servlet4demo.okhttpclient.dto.Greeting;
+import de.holisticon.servlet4demo.Greeting;
 import org.apache.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
@@ -9,7 +9,6 @@ import org.springframework.boot.CommandLineRunner;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.context.annotation.Bean;
-import org.springframework.http.client.OkHttp3ClientHttpRequestFactory;
 import org.springframework.web.client.RestTemplate;
 
 /**
