+++ demo-util/src/main/java/de/holisticon/servlet4demo/Greeting.java
@@ -1,8 +1,17 @@
-package de.holisticon.servlet4demo.okhttpclient.dto;
+package de.holisticon.servlet4demo;
 
 public class Greeting {
   private long id;
   private String content;
+
+  public Greeting() {
+  }
+
+  public Greeting(long id, String content) {
+
+    this.id = id;
+    this.content = content;
+  }
 
   public long getId() {
     return id;
