+++ demo-util/src/main/java/module-info.java
@@ -1,0 +1,4 @@
+module de.holisticon.servlet4demo {
+
+  exports de.holisticon.servlet4demo;
+}
