+++ pom.xml
@@ -8,7 +8,7 @@
     <version>0.0.1-SNAPSHOT</version>
     <packaging>pom</packaging>
 
-    <name>servlet4-demo</name>
+    <name>servlet4-demo-parent</name>
     <description>Demo project for Spring Boot</description>
 
     <parent>
@@ -24,6 +24,7 @@
         <module>server-glassfish</module>
         <module>server-undertow</module>
         <module>server-tomcat</module>
+        <module>demo-util</module>
     </modules>
 
     <licenses>
@@ -77,6 +78,11 @@
 
     <dependencyManagement>
         <dependencies>
+            <dependency>
+                <groupId>de.holisticon</groupId>
+                <artifactId>servlet4-demo-util</artifactId>
+                <version>0.0.1-SNAPSHOT</version>
+            </dependency>
             <dependency>
                 <groupId>org.springframework.boot</groupId>
                 <artifactId>spring-boot-starter-data-rest</artifactId>
