+++ server-jetty/src/main/java/module-info.java
@@ -1,23 +1,25 @@
 module de.holisticon.servlet4demo.serverjetty {
 
+  requires de.holisticon.servlet4demo;
+
   requires spring.core;
   requires spring.beans;
   requires spring.context;
   requires spring.aop;
   requires spring.web;
   requires spring.expression;
-
   requires spring.boot;
   requires spring.boot.autoconfigure;
+  requires spring.webmvc;
+
+  requires jetty.alpn.server;
+  requires jetty.http;
+  requires jetty.server;
   requires jetty.util;
   requires http2.common;
-  requires jetty.http;
-  requires jackson.annotations;
-  requires jetty.server;
-  requires javax.servlet.api;
   requires http2.server;
-  requires jetty.alpn.server;
-  requires spring.webmvc;
+  requires jackson.annotations;
+  requires javax.servlet.api;
   requires log4j.over.slf4j;
 
   exports de.holisticon.servlet4demo.serverjetty;
